+++ test/Fakes/CleanRegex/Match/Details/Group/TextGroup.php
@@ -102,7 +102,7 @@ class TextGroup implements Group
         throw $this->fail();
     }
 
-    public function orReturn($substitute)
+    public function or(string $substitute): string
     {
         throw $this->fail();
     }